Move RedundantByRefModifierInspectionTests for QuickFixes into relevant files under RD.QuickFixes
